Refactor testing harness to be more general, very WIP
