+++ tests/mod.rs
@@ -1,4 +1,4 @@
-mod common;
+mod harness;
 
 // Every subfolder with tests need to be defined as modules here, otherwise they won't be executed
 mod test_smoke;
